+++ Casks/focuslens.rb
@@ -1,6 +1,6 @@
 cask "focuslens" do
-  version "1.0.0"
-  sha256 "a95e70037b9dfeaa2c7b903e94e941abb08dd6b20f1ca0823169fc9704b0d3d8"
+  version "1.0.1"
+  sha256 "be98cf5b938a3f66aacba40ffc29fe6e4fa1f347852990b006e4837bc0ffa210"
 
   url "https://github.com/parththummar/FocusLens/releases/download/#{version}/FocusLens-#{version}.zip",
       verified: "github.com/parththummar/FocusLens/"